+++ mcctarget.hpp
@@ -8,7 +8,7 @@ struct MCCTarget
 	enum Type { Unknown, CC, Fixed };
 
 	Type type;
-	uint8_t value;
+	double value;
 	uint8_t channel;
 	uint8_t cc;
